fix rate limit test
